+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All Notable changes to `pickle` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## NEXT - YYYY-MM-DD
+## Working on RUN and better DI - 2017-05-13
 
 ### Added
-- Nothing
+- `pickle run` so you can run from the feature file ui or domain tests
+- better us of DI system
+
 
 ### Deprecated
 - Nothing